% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{coef.multiflexscan}
\alias{coef.multiflexscan}
\title{Extract selected cluster summaries}
\usage{
\method{coef}{multiflexscan}(object, ...)
}
\arguments{
\item{object}{A \code{multiflexscan} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A data frame of selected clusters (see \code{\link[=clusters]{clusters()}}).
}
\description{
Extract selected cluster summaries
}
