% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_setting}
\alias{get_setting}
\title{Analysis settings}
\usage{
get_setting(object)
}
\arguments{
\item{object}{A \code{multiflexscan} object.}
}
\value{
A list of analysis settings stored in the object.
}
\description{
Analysis settings
}
\seealso{
\code{\link[=nclusters]{nclusters()}}, \code{\link[=pvalue]{pvalue()}}
}
