% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiflexscan.R
\docType{package}
\name{multiflexscan-package}
\alias{multiflexscan-package}
\title{Detecting multiple spatial disease clusters using the information criterion
and scan statistic approach}
\description{
This package provides functions for detecting multiple spatial disease
clusters based on the information criterion and scan statistic approach
by Takahashi and Shimadzu (2020), using the flexible spatial scan statistic
(Tango and Takahashi, 2005) or Kulldorff's circular spatial scan statistic
(1997) implemented in the rflexscan package (Otani and Takahashi, 2021).
}
\references{
\itemize{
\item Takahashi K. and Shimadzu H. (2020). Detecting multiple spatial
disease clusters: information criterion and scan statistic approach.
International Journal of Health Geographics 19:33.
\doi{10.1186/s12942-020-00228-y}
\item Tango T. and Takahashi K. (2005). A flexibly shaped spatial scan
statistic for detecting clusters, International Journal of Health
Geographics 4:11.
\item Otani T. and Takahashi K. (2021). Flexible scan statistics for
detecting spatial disease clusters: The rflexscan R package.
Journal of Statistical Software 99:13.
\doi{10.18637/jss.v099.i13}
\item Takahashi K, Yokoyama T and Tango T. (2010). FleXScan v3.1: Software
for the Flexible Scan Statistic. National Institute of Public Health, Japan,
\url{https://sites.google.com/site/flexscansoftware/home}.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tkhrotn/multiflexscan}
  \item Report bugs at \url{https://github.com/tkhrotn/multiflexscan/issues}
}

}
\author{
\strong{Maintainer}: Takahiro Otani \email{t.otani@aichi-cc.jp} (\href{https://orcid.org/0000-0002-4007-7817}{ORCID})

Authors:
\itemize{
  \item Takahiro Otani \email{t.otani@aichi-cc.jp} (\href{https://orcid.org/0000-0002-4007-7817}{ORCID})
  \item Kunihiko Takahashi
}

}
