% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{nclusters}
\alias{nclusters}
\title{Number of clusters selected by the information criterion}
\usage{
nclusters(object)
}
\arguments{
\item{object}{A \code{multiflexscan} object.}
}
\value{
Integer number of selected clusters.
}
\description{
Number of clusters selected by the information criterion
}
\seealso{
\code{\link[=pvalue]{pvalue()}}, \code{\link[=clusters]{clusters()}}, \code{\link[=get_setting]{get_setting()}}
}
