% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{nobs.multiflexscan}
\alias{nobs.multiflexscan}
\title{Number of regions in the analysis}
\usage{
\method{nobs}{multiflexscan}(object, ...)
}
\arguments{
\item{object}{A \code{multiflexscan} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Number of regions (rows) in the input data.
}
\description{
Number of regions in the analysis
}
