% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiflexscan.R
\name{plot.multiflexscan}
\alias{plot.multiflexscan}
\title{Plot method for the \code{multiflexscan} object. The neighborhood graph is
drawn with \pkg{igraph}, following \code{plot.rflexscan()}.}
\usage{
\method{plot}{multiflexscan}(
  x,
  rank = 1:x$nclust,
  pval = 1,
  vertexsize = max(x$input$coordinates[, 1]) - min(x$input$coordinates[, 1]),
  xlab = colnames(x$input$coordinates)[1],
  ylab = colnames(x$input$coordinates)[2],
  xlim = c(min(x$input$coordinates[, 1]), max(x$input$coordinates[, 1])),
  ylim = c(min(x$input$coordinates[, 2]), max(x$input$coordinates[, 2])),
  col = palette(),
  frame_color = "gray40",
  vertex_color = "white",
  ...
)
}
\arguments{
\item{x}{A \code{multiflexscan} object.}

\item{rank}{Integer vector specifying the clusters to highlight.}

\item{pval}{Maximum cluster-level p-value for a candidate cluster to be highlighted.}

\item{vertexsize}{Size of region centroids in the igraph plot.}

\item{xlab}{Label of the x-axis.}

\item{ylab}{Label of the y-axis.}

\item{xlim}{Limits of the x-axis.}

\item{ylim}{Limits of the y-axis.}

\item{col}{Colors used for highlighted clusters.}

\item{frame_color}{Color used for neighborhood edges and vertex frames not highlighted as
clusters.}

\item{vertex_color}{Color used for region centroids that are not highlighted as clusters.}

\item{...}{Additional arguments passed to \code{igraph::plot.igraph()}.}
}
\value{
The input object, invisibly.
}
\description{
Plot method for the \code{multiflexscan} object. The neighborhood graph is
drawn with \pkg{igraph}, following \code{plot.rflexscan()}.
}
\seealso{
\link{multiflexscan}, \link{choropleth}, \link[rflexscan:plot.rflexscan]{rflexscan::plot.rflexscan}
}
