% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiflexscan.R
\name{print.multiflexscan}
\alias{print.multiflexscan}
\title{Print multiflexscan object}
\usage{
\method{print}{multiflexscan}(x, ...)
}
\arguments{
\item{x}{A \code{multiflexscan} object to be printed.}

\item{...}{Ignored.}
}
\value{
The input object, invisibly.
}
\description{
Print method for \code{multiflexscan} objects.
}
\seealso{
\link{multiflexscan}, \link{summary.multiflexscan}
}
