% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiflexscan.R
\name{print.summary.multiflexscan}
\alias{print.summary.multiflexscan}
\title{Print summary of multiflexscan results}
\usage{
\method{print}{summary.multiflexscan}(x, ...)
}
\arguments{
\item{x}{A \code{summary.multiflexscan} object to be printed.}

\item{...}{Ignored.}
}
\value{
The input object, invisibly.
}
\description{
Print summary of \code{multiflexscan} results to the terminal.
}
\seealso{
\link{multiflexscan}, \link{summary.multiflexscan}
}
