% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{pvalue}
\alias{pvalue}
\title{Overall Monte Carlo p-value}
\usage{
pvalue(object)
}
\arguments{
\item{object}{A \code{multiflexscan} object.}
}
\value{
Numeric overall p-value for the multiple-cluster detection test.
}
\description{
Overall Monte Carlo p-value
}
\seealso{
\code{\link[=nclusters]{nclusters()}}, \code{\link[=clusters]{clusters()}}
}
